Refactor: Modularize the main application script
The main `app.js` file has been refactored from a single monolithic script into a set of smaller, more maintainable modules. Each module is responsible for a specific piece of functionality, such as CEP integration, UI management, icon handling, or error recovery.

This modularization improves the overall code structure, making it easier to read, debug, and extend in the future. The original `app.js` now serves as a clean entry point that initializes the application by importing the necessary modules.

The new file structure is as follows:
- `app.js`: The main application entry point.
- `panel.js`: Contains the core `MaterialIconsPanel` class and main application logic.
- `cep.js`: Handles all communication with the Adobe CEP environment.
- `icons.js`: Manages icon loading, filtering, and rendering.
- `ui.js`: Provides UI-related functionalities like loading indicators and notifications.
- `dom.js`: Manages DOM readiness checks.
- `debug.js`: Contains the debugging and logging utility.
- `errorRecovery.js`: Provides error handling and retry mechanisms.
